frontend/requirements.txt
streamlit
requests
pandas
numpy
yfinance
plotly
python-dotenv